feat: add auth and query movescription
